Added fitting methods to bosonic environments. Fitting code moved into utilities module.
