+++ qutip/core/environment.py
@@ -14,7 +14,6 @@ __all__ = ['BosonicEnvironment',
 
 import abc
 import enum
-from time import time
 from typing import Any, Callable
 import warnings
 
@@ -22,7 +21,6 @@ import numpy as np
 from numpy.typing import ArrayLike
 from scipy.integrate import quad_vec
 from scipy.interpolate import interp1d
-from scipy.interpolate import InterpolatedUnivariateSpline
 
 try:
     from mpmath import mp
@@ -30,9 +28,7 @@ try:
 except ModuleNotFoundError:
     _mpmath_available = False
 
-from ..utilities import n_thermal
-from .fit_utils import (_run_fit, _gen_summary,
-                        _two_column_summary, aaa, filter_poles)
+from ..utilities import (n_thermal, CorrelationFitter, SpectralFitter)
 
 
 # TODO fermionic environment
@@ -62,6 +58,14 @@ class BosonicEnvironment(abc.ABC):
     def __init__(self, T: float = None, *, tag: Any = None):
         self.T = T
         self.tag = tag
+
+        # TODO should these be available always?
+        # TODO what if T is not set?
+        # TODO add AAA fitting in later PR
+        self._avail_approximators = {
+            'correlation_function_fit': self._fit_correlation_function,
+            'underdamped_fit': self._fit_spectral_density,
+        }
 
     @abc.abstractmethod
     def spectral_density(self, w: float | ArrayLike):
@@ -97,6 +101,57 @@ class BosonicEnvironment(abc.ABC):
         else:
             S = 2 * np.heaviside(w, 0) * self.spectral_density(w)
         return S
+
+    def exponential_approximation(self, method, **kwargs):
+        # TODO documentation
+        # TODO is this the right way of doing it?
+        approximator = self._avail_approximators.get(method, None)
+        if approximator is None:
+            raise ValueError(f"Unknown approximation method: {method}")
+        return approximator(**kwargs)
+
+    def _fit_correlation_function(self, tlist, Nr, Ni, full_ansatz=False):
+        """
+        Generates a reservoir from the correlation function
+
+        Parameters
+        ----------
+        tlist: obj:`np.ndarray`
+            The time range on which to perform the fit
+        Nr: int
+            The number of modes to use for the fit of the real part 
+        Ni: int
+            The number of modes to use for the fit of the real part
+        full_ansatz: bool
+            Whether to use a fit of the imaginary and real parts that is 
+            complex
+
+        Returns
+        -------
+        A bosonic reservoir
+        """
+        fitter = CorrelationFitter(self.T, tlist, self.correlation_function)
+        return fitter.get_fit(Nr=Nr, Ni=Ni, full_ansatz=full_ansatz)
+
+    def _fit_spectral_density(self, wlist, N, Nk):
+        """
+        Generates a reservoir from the spectral density
+
+        Parameters
+        ----------
+        wlist: obj:`np.ndarray`
+            The frequency range on which to perform the fit
+        N: int
+            The number of modes to use for the fit
+        Nk: int
+            The number of exponents to use in each mode
+
+        Returns
+        -------
+        A bosonic reservoir
+        """
+        fitter = SpectralFitter(self.T, wlist, self.spectral_density)
+        return fitter.get_fit(N=N, Nk=Nk)
 
     # TODO: I thought these temperatures can be `None`. In what cases is that ok?
     @classmethod
@@ -930,594 +985,3 @@ def _fft(fun, t0=10, dt=1e-5):
     sorted_indices = np.argsort(w)
     zz = interp1d(w[sorted_indices], g[sorted_indices])
     return zz
-
-
-
-
-
-# --- old code ---
-
-
-class ApproximatedBosonicBath:
-    """
-    This class allows to construct a reservoir from the correlation function,
-    power spectrum or spectral density.
-    """
-    @classmethod
-    def from_sd(self, bath, N, Nk, x, Q):
-        """
-        Generates a reservoir from the spectral density
-
-        Parameters
-        ----------
-        bath: obj:`BosonicReservoir`
-            The reservoir we want to approximate
-        N: int
-            The number of modes to use for the fit
-        Nk: int
-            The number of exponents to use in each mode
-        x: obj:`np.ndarray`
-            The range on which to perform the fit
-        Q: obj:`Qobj`
-            The coupling operator to the bath
-
-        Returns
-        -------
-        A bosonic reservoir
-        """
-        cls = SpectralFitter(bath.T, Q, x, bath.spectral_density)
-        cls.get_fit(N=N, Nk=Nk)
-        return cls
-
-    @classmethod
-    def from_ps(self, bath, x, tol, max_exponents, Q):
-        """
-        Generates a reservoir from the power spectrum
-
-        Parameters
-        ----------
-        bath: obj:`BosonicReservoir`
-            The reservoir we want to approximate
-        tol: float
-            The desired error tolerance
-        max_exponents: int
-            The maximum number of exponents allowed
-        x: obj:`np.ndarray`
-            The range on which to perform the fit
-        Q: obj:`Qobj`
-            The coupling operator to the bath
-
-        Returns
-        -------
-        A bosonic reservoir
-        """
-        r, pol, res, zer, _ = aaa(bath.power_spectrum, x,
-                                  tol=tol,
-                                  max_iter=max_exponents*2)
-        new_pols, new_res = filter_poles(pol, res)
-        ckAR, ckAI = np.real(-1j*new_res), np.imag(-1j*new_res)
-        vkAR, vkAI = np.real(1j*new_pols), np.imag(1j*new_pols)
-        cls = ExponentialBosonicBath(
-            Q=Q, ck_real=ckAR, vk_real=vkAR + 1j * vkAI, ck_imag=ckAI,
-            vk_imag=vkAR + 1j * vkAI, T=bath.T)
-        return cls
-
-    @classmethod
-    def from_cf(self, Q, x, bath, Nr, Ni, full_ansatz=False):
-        """
-        Generates a reservoir from the correlation function
-
-        Parameters
-        ----------
-        bath: obj:`BosonicReservoir`
-            The reservoir we want to approximate
-        Nr: int
-            The number of modes to use for the fit of the real part 
-        Ni: int
-            The number of modes to use for the fit of the real part 
-        x: obj:`np.ndarray`
-            The range on which to perform the fit
-        Q: obj:`Qobj`
-            The coupling operator to the bath
-        full_ansatz: bool
-            Whether to use a fit of the imaginary and real parts that is 
-            complex
-        Returns
-        -------
-        A bosonic reservoir
-        """
-        cls = CorrelationFitter(Q, bath.T, x, bath.correlation_function)
-        cls.get_fit(
-            Nr=Nr,
-            Ni=Ni,
-            full_ansatz=full_ansatz)
-        return cls
-
-
-class SpectralFitter(ExponentialBosonicBath):
-    """
-    A helper class for constructing a Bosonic bath from a fit of the spectral
-    density with a sum of underdamped modes.
-
-    Parameters
-    ----------
-    Q : :obj:`.Qobj`
-        Operator describing the coupling between system and bath.
-
-    T : float
-        Bath temperature.
-
-    w : :obj:`np.array.`
-        The range on which to perform the fit, it is recommended that it covers
-        at least twice the cutoff frequency of the desired spectral density.
-
-    J : :obj:`np.array.` or callable
-        The spectral density to be fitted as an array or function.
-    """
-
-    def __init__(self, T, Q, w, J):
-        self.Q = Q
-        self.T = T
-        self.fitinfo = None
-        self.set_spectral_density(w, J)
-
-    def set_spectral_density(self, w, J):
-        """
-        Sets the spectral density to be fitted. It may be provided either as an
-        array of function values or as a python function. For internal reasons,
-        it will then be interpolated or discretized as necessary.
-        """
-
-        if callable(J):
-            self._w = w
-            self._J_array = J(w)
-            self._J_fun = J
-        else:
-            self._w = w
-            self._J_array = J
-            self._J_fun = InterpolatedUnivariateSpline(w, J)
-
-    @classmethod
-    def _meier_tannor_SD(cls, w, a, b, c):
-        r"""
-        Underdamped spectral density used for fitting in Meier-Tannor form
-        (see Eq. 38 in the BoFiN paper, DOI: 10.1103/PhysRevResearch.5.013181)
-        or the get_fit method.
-
-        Parameters
-        ----------
-        w : :obj:`np.array.`
-            The frequency of the spectral density
-        a : :obj:`np.array.`
-            Array of coupling constants ($\alpha_i^2$)
-        b : :obj:`np.array.`
-            Array of cutoff parameters ($\Gamma'_i$)
-        c : :obj:`np.array.`
-            Array of resonant frequencies ($\Omega_i$)
-        """
-
-        return sum((2 * ai * bi * w
-                    / ((w + ci) ** 2 + bi ** 2)
-                    / ((w - ci) ** 2 + bi ** 2))
-                   for ai, bi, ci in zip(a, b, c))
-
-    def get_fit(
-        self,
-        N=None,
-        Nk=1,
-        final_rmse=5e-6,
-        lower=None,
-        upper=None,
-        sigma=None,
-        guesses=None,
-    ):
-        r"""
-        Provides a fit to the spectral density with N underdamped oscillator
-        baths. N can be determined automatically based on reducing the
-        normalized root mean squared error below a certain threshold.
-
-        Parameters
-        ----------
-        N : optional, int
-            Number of underdamped oscillators to use.
-            If set to None, it is determined automatically.
-        Nk : optional, int
-            Number of exponential terms used to approximate the bath
-            correlation functions, defaults to 1. To approximate the
-            correlation function the number of exponents grow as the
-            Desired normalized root mean squared error. Defaults to
-            :math:`5\times10^{-6}`. Only used if N is set to None.
-            Desired normalized root mean squared error. Defaults to
-            Lower bounds on the parameters for the fit. A list of size 3,
-            containing the lower bounds for :math:`a_i` (coupling constants),
-            :math:`b_i` (cutoff frequencies) and :math:`c_i`
-            (resonant frequencies) in the following fit function:
-
-            .. math::
-                J(\omega) = \sum_{i=1}^{k} \frac{2 a_{i} b_{i} \omega
-                }{\left(\left( \omega + c_{i}\right)^{2} + b_{i}^{2}\right)
-                \left(\left( \omega - c_{i}\right)^{2} + b_{i}^{2} \right)}
-
-            The lower bounds are considered to be the same for all N modes.
-            For example,
-
-            lower=[0,-1,2]
-
-            would bound the coupling to be bigger than 0, the cutoff frequency
-            to be higher than 1, and the central frequency to be bigger than 2
-
-        upper : list
-            Upper bounds on the parameters for the fit, the structure is the
-            same as the lower keyword.
-        sigma : float
-            Uncertainty in the data considered for the fit, all data points are
-            considered to have the same uncertainty.
-        guesses : list
-            Initial guesses for the parameters. Same structure as lower and
-            upper.
-
-        Note: If one of lower, upper, sigma, guesses is None, all are discarded
-
-        Returns
-        -------
-        1. A Bosonic Bath created with the fit parameters for the original
-          spectral density function (that was provided or interpolated)
-        2. A dictionary containing the following information about the fit:
-            * fit_time:
-                The time the fit took in seconds.
-            * rsme:
-                Normalized mean squared error obtained in the fit.
-            * N:
-                The number of terms used for the fit.
-            * params:
-                The fitted parameters (3N parameters), it contains three lists
-                one for each parameter, each list containing N terms.
-            * Nk:
-                The number of exponents used to construct the bosonic bath.
-            * summary:
-                A string that summarizes the information of the fit.
-        """
-
-        start = time()
-        rmse, params = _run_fit(
-            SpectralFitter._meier_tannor_SD, self._J_array, self._w,
-            final_rmse, default_guess_scenario="Spectral Density", N=N,
-            sigma=sigma, guesses=guesses, lower=lower, upper=upper)
-        end = time()
-
-        fit_time = end - start
-        spec_n = len(params[0])
-        self._generate_bath(params, Nk)
-        summary = _gen_summary(
-            fit_time, rmse, N, "The Spectral Density", params)
-        fitInfo = {
-            "fit_time": fit_time, "rmse": rmse, "N": spec_n, "params": params,
-            "Nk": Nk, "summary": summary}
-        self.fitinfo = fitInfo
-
-    def _generate_bath(self, params, Nk):
-        """
-        Obtains the bath exponents from the list of fit parameters. Some
-        transformations are done, to reverse the ones in the UnderDampedBath.
-        They are done to change the spectral density from eq. 38 to eq. 16
-        of the BoFiN paper and vice-versa.
-
-        Parameters
-        ----------
-        params: list
-            The parameters obtained from the fit.
-
-        Returns
-        -------
-            A Bosonic Bath created with the fit parameters.
-        """
-
-        lam, gamma, w0 = params
-        w0 = np.array(
-            [
-                np.sqrt((w0[i] + 0j) ** 2 + (gamma[i] + 0j / 2) ** 2)
-                for i in range(len(w0))
-            ]
-        )
-        lam = np.sqrt(
-            lam + 0j
-        )
-        # both w0, and lam modifications are needed to input the
-        # right value of the fit into the Underdamped bath
-        ckAR = []
-        vkAR = []
-        ckAI = []
-        vkAI = []
-
-        for lamt, Gamma, Om in zip(lam, gamma, w0):
-#            coeffs = UnderDampedBath._matsubara_params(
-#                lamt, 2 * Gamma, Om + 0j, self.T, Nk)
-#            ckAR.extend(coeffs[0])
-#            vkAR.extend(coeffs[1])
-#            ckAI.extend(coeffs[2])
-#            vkAI.extend(coeffs[3])
-            ... # TODO
-
-        super().__init__(self.Q, ckAR, vkAR, ckAI, vkAI, T=self.T)
-
-
-class CorrelationFitter(ExponentialBosonicBath):
-    """
-    A helper class for constructing a Bosonic bath from a fit of the
-    correlation function with exponential terms.
-
-    Parameters
-    ----------
-    Q : :obj:`.Qobj`
-        Operator describing the coupling between system and bath.
-    T : float
-        Temperature of the bath.
-    t : :obj:`np.array.`
-        The range which to perform the fit.
-    C : :obj:`np.array.` or callable
-        The correlation function to be fitted as an array or function.
-    """
-
-    def __init__(self, Q, T, t, C):
-        self.Q = Q
-        self.T = T
-        self.fitinfo = None
-        self.set_correlation_function(t, C)
-
-    def set_correlation_function(self, t, C):
-        """
-        This function creates a discretized version of the correlation function
-        if the correlation function is provided, and a function if
-        an array is provided.
-
-        The array is needed to run the least squares algorithm, while the
-        the function is used to assign a correlation function to the bosonic
-        bath object.
-        """
-        if callable(C):
-            self._t = t
-            self._C_array = C(t)
-            self._C_fun = C
-        else:
-            self._t = t
-            self._C_array = C
-            _C_fun_r = InterpolatedUnivariateSpline(t, np.real(C))
-            _C_fun_i = InterpolatedUnivariateSpline(t, np.imag(C))
-            self._C_fun = lambda t: _C_fun_r(t) + 1j * _C_fun_i(t)
-
-    def _corr_approx(self, t, a, b, c, d=0):
-        r"""
-        This is the form of the correlation function to be used for fitting.
-
-        Parameters
-        ----------
-        t : :obj:`np.array.` or float
-            The times at which to evaluates the correlation function.
-        a : list or :obj:`np.array.`
-            A list describing the  real part amplitude of the correlation
-            approximation.
-        b : list or :obj:`np.array.`
-            A list describing the decay of the correlation approximation.
-        c : list or :obj:`np.array.`
-            A list describing the oscillations of the correlation
-            approximation.
-        d:  A list describing the imaginary part amplitude of the correlation
-            approximation, only used if the user selects if the full_ansatz
-            flag from get_fit is True.
-        """
-
-        a = np.array(a)
-        b = np.array(b)
-        c = np.array(c)
-        d = np.array(d)
-        if (d == 0).all():
-            d = np.zeros(a.shape)
-
-        return np.sum(
-            (a[:, None]+1j*d[:, None]) * np.exp(b[:, None] * t[None, :]) *
-            np.exp(1j*c[:, None] * t[None, :]),
-            axis=0,
-        )
-
-    def get_fit(
-        self,
-        Nr=None,
-        Ni=None,
-        final_rmse=2e-5,
-        lower=None,
-        upper=None,
-        sigma=None,
-        guesses=None,
-        full_ansatz=False
-    ):
-        r"""
-        Fit the correlation function with Ni exponential terms
-        for the imaginary part of the correlation function and Nr for the real.
-        If no number of terms is provided, this function determines the number
-        of exponents based on reducing the normalized root mean squared
-        error below a certain threshold.
-
-        Parameters
-        ----------
-        Nr : optional, int
-            Number of exponents to use for the real part.
-            If set to None it is determined automatically.
-        Ni : optional, int
-            Number of exponents terms to use for the imaginary part.
-            If set to None it is found automatically.
-        final_rmse : float
-            Desired normalized root mean squared error. Only used if Ni or Nr
-            are not specified.
-        lower : list
-            lower bounds on the parameters for the fit. A list of size 4 when
-            full_ansatz is True and of size 3 when it is false,each value
-            represents the lower bound for each parameter.
-
-            The first and last terms describe the real and imaginary parts of
-            the amplitude, the second the decay rate, and the third one the
-            oscillation frequency. The lower bounds are considered to be
-            the same for all Nr and Ni exponents. for example
-
-            lower=[0,-1,1,1]
-
-            would bound the real part of the amplitude to be bigger than 0,
-            the decay rate to be higher than -1, and the oscillation frequency
-            to be bigger than 1, and the imaginary part of the amplitude to
-            be greater than 1
-        upper : list
-            upper bounds on the parameters for the fit, the structure is the
-            same as the lower keyword.
-        sigma : float
-            uncertainty in the data considered for the fit, all data points are
-            considered to have the same uncertainty.
-        guesses : list
-            Initial guesses for the parameters. Same structure as lower and
-            upper.
-        full_ansatz : bool
-            Indicates whether to use the function
-
-            .. math::
-                C(t)= \sum_{k}a_{k}e^{-b_{k} t}e^{i c_{k} t}
-
-            for the fitting of the correlation function (when False, the
-            default value)  this function gives us
-            faster fits,usually it is not needed to tweek
-            guesses, sigma, upper and lower as defaults work for most
-            situations.  When set to True one uses the function
-
-            .. math::
-                C(t)= \sum_{k}(a_{k}+i d_{k})e^{-b_{k} t}e^{i c_{k} t}
-
-            Unfortunately this gives us significantly slower fits and some
-            tunning of the guesses,sigma, upper and lower are usually needed.
-            On the other hand, it can lead to better fits with lesser exponents
-            specially for anomalous spectral densities such that
-            $Im(C(0))\neq 0$. When using this with default values if the fit
-            takes too long you should input guesses, lower and upper bounds,
-            if you are not sure what to set them to it is useful to use the
-            output of fitting with the other option as guesses for the fit.
-
-
-
-        Note: If one of lower, upper, sigma, guesses is None, all are discarded
-
-        Returns
-        -------
-        1. A Bosonic Bath created with the fit parameters from the original
-          correlation function (that was provided or interpolated).
-        2. A dictionary containing the following information about the fit:
-            * Nr :
-                The number of terms used to fit the real part of the
-                correlation function.
-            * Ni :
-                The number of terms used to fit the imaginary part of the
-                correlation function.
-            * fit_time_real :
-                The time the fit of the real part of the correlation function
-                took in seconds.
-            * fit_time_imag :
-                The time the fit of the imaginary part of the correlation
-                function took in seconds.
-            * rsme_real :
-                Normalized mean squared error obtained in the fit of the real
-                part of the correlation function.
-            * rsme_imag :
-                Normalized mean squared error obtained in the fit of the
-                imaginary part of the correlation function.
-            * params_real :
-                The fitted parameters (3N parameters) for the real part of the
-                correlation function, it contains three lists one for each
-                parameter, each list containing N terms.
-            * params_imag :
-                The fitted parameters (3N parameters) for the imaginary part
-                of the correlation function, it contains three lists one for
-                each parameter, each list containing N terms.
-            * summary :
-                A string that summarizes the information about the fit.
-            """
-        if full_ansatz:
-            num_params = 4
-        else:
-            num_params = 3
-        # Fit real part
-        start_real = time()
-        rmse_real, params_real = _run_fit(
-            lambda *args: np.real(self._corr_approx(*args)),
-            y=np.real(self._C_array), x=self._t, final_rmse=final_rmse,
-            default_guess_scenario="correlation_real", N=Nr, sigma=sigma,
-            guesses=guesses, lower=lower, upper=upper, n=num_params)
-        end_real = time()
-
-        # Fit imaginary part
-        start_imag = time()
-        rmse_imag, params_imag = _run_fit(
-            lambda *args: np.imag(self._corr_approx(*args)),
-            y=np.imag(self._C_array), x=self._t, final_rmse=final_rmse,
-            default_guess_scenario="correlation_imag", N=Ni, sigma=sigma,
-            guesses=guesses, lower=lower, upper=upper, n=num_params)
-        end_imag = time()
-
-        # Calculate Fit Times
-        fit_time_real = end_real - start_real
-        fit_time_imag = end_imag - start_imag
-
-        # Generate summary
-        Nr = len(params_real[0])
-        Ni = len(params_imag[0])
-        full_summary = _two_column_summary(
-            params_real, params_imag, fit_time_real, fit_time_imag, Nr, Ni,
-            rmse_imag, rmse_real, n=num_params)
-
-        fitInfo = {"Nr": Nr, "Ni": Ni,
-                   "fit_time_real": fit_time_real,
-                   "fit_time_imag": fit_time_imag,
-                   "rmse_real": rmse_real, "rmse_imag": rmse_imag,
-                   "params_real": params_real,
-                   "params_imag": params_imag, "summary": full_summary}
-        self._generate_bath(params_real, params_imag, n=num_params)
-        self.fitinfo = fitInfo
-
-    def _generate_bath(self, params_real, params_imag, n=3):
-        """
-        Calculate the Matsubara coefficients and frequencies for the
-        fitted underdamped oscillators and generate the corresponding bosonic
-        bath.
-
-        Parameters
-        ----------
-        params_real : :obj:`np.array.`
-            array of shape (N,3) where N is the number of fitted terms
-            for the real part.
-        params_imag : np.imag
-            array of shape (N,3) where N is the number of fitted terms
-            for the imaginary part.
-
-        Returns
-        -------
-        A bosonic Bath constructed from the fitted exponents.
-        """
-        if n == 4:
-            a, b, c, d = params_real
-            a2, b2, c2, d2 = params_imag
-        else:
-            a, b, c = params_real
-            a2, b2, c2 = params_imag
-            d = np.zeros(a.shape, dtype=int)
-            d2 = np.zeros(a2.shape, dtype=int)
-
-        # the 0.5 is from the cosine
-        ckAR = [(x + 1j*y)*0.5 for x, y in zip(a, d)]
-        # extend the list with the complex conjugates:
-        ckAR.extend(np.conjugate(ckAR))
-        vkAR = [-x - 1.0j * y for x, y in zip(b, c)]
-        vkAR.extend([-x + 1.0j * y for x, y in zip(b, c)])
-
-        # the 0.5 is from the sine
-        ckAI = [-1j*(x + 1j*y)*0.5 for x, y in zip(a2, d2)]
-
-        # extend the list with the complex conjugates:
-        ckAI.extend(np.conjugate(ckAI))
-        vkAI = [-x - 1.0j * y for x, y in zip(b2, c2)]
-        vkAI.extend([-x + 1.0j * y for x, y in zip(b2, c2)])
-
-        super().__init__(self.Q, ckAR, vkAR, ckAI, vkAI, T=self.T)
